// src/Framework/SPPC.Tadbir.Web.New/ClientApp/src/app/model/accountItemBrief.ts
// ------------------------------------------------------------------------------
// <auto-generated>
//     This code was generated by a tool.
//     Runtime Version: 1.1.449
//     Template Version: 1.0
//     Generation Date: 02/02/2019 02:26:17 ب.ظ
//  
//     Changes to this file may cause incorrect behavior and will be lost if
//     the code is regenerated.
// </auto-generated>
// ------------------------------------------------------------------------------


export interface AccountItemBrief {
    id: number;
    name: string;
    code: string;
    fullCode: string;
    level: number;
    childCount: number;
    parentId?: number;
    isSelected: boolean;
}